refactor: add ArticleMetadata component
