+++ src/App.tsx
@@ -1,15 +1,7 @@
-import { useState } from "react";
 import article from "./data.json";
-import { SharePopover } from "./components/SharePopover";
-import { ShareBar } from "./components/ShareBar";
+import { ArticleMetadata } from "./components/ArticleMetadata";
 
 function App() {
-  const [shareOpen, setShareOpen] = useState(false);
-
-  function toggleShare() {
-    setShareOpen((prev) => !prev);
-  }
-
   return (
     <main className="min-h-screen grid justify-center items-start lg:place-content-center">
       <article className="my-[74px] mx-6 max-w-[520px] lg:max-w-[730px] flex flex-col lg:flex-row gap-8 lg:gap-0 rounded-[10px] bg-white shadow-[0px_40px_40px_-10px_rgba(201,213,225,0.50)]">
@@ -21,38 +13,13 @@ function App() {
           />
         </header>
         <section className="space-y-3 lg:pt-8">
-          <h2 className="px-8 lg:px-10 font-bold lg:text-xl">{article.title}</h2>
+          <h2 className="px-8 lg:px-10 font-bold lg:text-xl">
+            {article.title}
+          </h2>
           <p className="px-8 lg:px-10 font-medium text-sm text-[#6E8098]">
             {article.content}
           </p>
-
-          <footer className="px-8 lg:px-10 py-5 lg:pb-10 relative">
-            <div className="flex items-center gap-4">
-              <img
-                src={article.author.image.url}
-                alt={article.author.image.alt}
-                className="w-10 h-10 rounded-full"
-              />
-              <div>
-                <h2 className="text-sm font-bold">{article.author.name}</h2>
-                <h3 className="text-sm text-light-gray">
-                  {article.author.date}
-                </h3>
-              </div>
-
-              <div className="relative ml-auto">
-                <button className="rounded-full shrink-0 align-middle" onClick={toggleShare}>
-                  {shareOpen ? (
-                    <img src="/assets/icon-share-active.svg" alt="Share icon" />
-                  ) : (
-                    <img src="/assets/icon-share.svg" alt="Share icon" />
-                  )}
-                </button>
-                <SharePopover open={shareOpen} />
-              </div>
-            </div>
-            <ShareBar open={shareOpen} handleClick={toggleShare} />
-          </footer>
+          <ArticleMetadata author={article.author} />
         </section>
       </article>
     </main>
